uint8_t, not void

+++ Firmware/ChameleonMini/Terminal/Terminal.c
@@ -43,7 +43,7 @@ void TerminalSendHEXBlock(const void* Buffer, uint16_t ByteCount)
     char sendArray[4];
     for (uint16_t i = 0; i < ByteCount; i++)
     {
-        sprintf(sendArray, "%02X", Buffer[i]);
+        sprintf(sendArray, "%02X", (uint8_t)Buffer[i]);
         TerminalSendString(sendArray);
     }
 }
